Struct diem_json_rpc_client::async_client::Client

pub struct Client<R> {
    pub http_client: Arc<dyn HttpClient>,
    pub retry: R,
}

Fields
http_client: Arc<dyn HttpClient>retry: R
Implementations

impl<R: RetryStrategy> Client<R>

pub fn from_url<T: IntoUrl>(server_url: T, retry: R) -> Result<Self, Error>

pub fn from_url_list<T: IntoUrl>(
server_urls: Vec<T>,
num_parallel_requests: usize,
retry: R
) -> Result<Self, String>

pub async fn wait_for_signed_transaction<'_, '_>(
&'_ self,
txn: &'_ SignedTransaction,
timeout: Option<Duration>,
delay: Option<Duration>
) -> Result<Response<Transaction>, WaitForTransactionError>

pub async fn wait_for_transaction<'_, '_, '_>(
&'_ self,
address: &'_ AccountAddress,
seq: u64,
expiration_time_secs: u64,
txn_hash: &'_ str,
timeout: Option<Duration>,
delay: Option<Duration>
) -> Result<Response<Transaction>, WaitForTransactionError>

pub fn last_known_state(&self) -> Option<State>

pub async fn get_metadata<'_>(&'_ self) -> Result<Response<Metadata>, Error>

pub async fn get_metadata_by_version<'_>(
&'_ self,
version: u64
) -> Result<Response<Metadata>, Error>

pub async fn get_account<'_, '_>(
&'_ self,
address: &'_ AccountAddress
) -> Result<Response<Option<Account>>, Error>

pub async fn get_account_transaction<'_, '_>(
&'_ self,
address: &'_ AccountAddress,
seq: u64,
include_events: bool
) -> Result<Response<Option<Transaction>>, Error>

pub async fn get_account_transactions<'_, '_>(
&'_ self,
address: &'_ AccountAddress,
start_seq: u64,
limit: u64,
include_events: bool
) -> Result<Response<Vec<Transaction>>, Error>

pub async fn get_transactions<'_>(
&'_ self,
start_seq: u64,
limit: u64,
include_events: bool
) -> Result<Response<Vec<Transaction>>, Error>

pub async fn get_events<'_, '_>(
&'_ self,
key: &'_ str,
start_seq: u64,
limit: u64
) -> Result<Response<Vec<Event>>, Error>

pub async fn get_events_with_proofs<'_, '_>(
&'_ self,
key: &'_ str,
start_seq: u64,
limit: u64
) -> Result<Response<Vec<Event>>, Error>

pub async fn get_currencies<'_>(
&'_ self
) -> Result<Response<Vec<CurrencyInfo>>, Error>

pub async fn get_state_proof<'_>(
&'_ self,
from_version: u64
) -> Result<Response<StateProof>, Error>

pub async fn get_account_state_with_proof<'_, '_>(
&'_ self,
address: &'_ AccountAddress,
from_version: Option<u64>,
to_version: Option<u64>
) -> Result<Response<AccountStateWithProof>, Error>

pub async fn submit<'_, '_>(
&'_ self,
txn: &'_ SignedTransaction
) -> Result<Response<()>, Error>

pub async fn send<T: DeserializeOwned, '_>(
&'_ self,
request: Request
) -> Result<Response<T>, Error>

pub async fn send_opt<T: DeserializeOwned, '_>(
&'_ self,
request: Request
) -> Result<Response<Option<T>>, Error>

pub async fn send_with_retry<RS: RetryStrategy, '_, '_, '_>(
&'_ self,
request: &'_ Request,
retry: &'_ RS
) -> Result<JsonRpcResponse, Error>

pub async fn batch_send<'_>(
&'_ self,
requests: Vec<Request>
) -> Result<Vec<JsonRpcResponse>, Error>

Batch requests into one JSON-RPC batch request. To keep interface simple, this method returns error when any error occurs. When batch responses contain partial success response, we will return the first error it hit.

Caller can convert JsonRpcResponse into Response<T> in order of given requests:

use diem_json_rpc_client::async_client::{
    types as jsonrpc, Client, Error, Request, Response, Retry,
};
use std::convert::{TryInto};

let client = Client::from_url("http://testnet.diem.com/v1", Retry::default()).unwrap();
let mut res = client
    .batch_send(vec![Request::get_metadata(), Request::get_currencies()])
    .await?;

let metadata: Response<jsonrpc::Metadata> = res
    .remove(0)
    .try_into()?;

let currencies: Response<jsonrpc::CurrencyInfo> = res
    .remove(0)
    .try_into()?;

For get_account and get_account_transaction request, the above convert will return result not found error instead of None.

This function calls to batch_send_with_retry, when you batch submit transactions, to avoid re-submit errors, it is better to use batch_send_without_retry and handle errors by yourself.

pub async fn batch_send_with_retry<RS: RetryStrategy, '_, '_, '_>(
&'_ self,
requests: &'_ [Request],
retry: &'_ RS
) -> Result<Vec<JsonRpcResponse>, Error>